# core/vm, go.mod: update uint256 and use faster method to write to memory (#30868)