@minh-bq 2024-12-19
On CI the PR shows slight improvement in more general benchmarks

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm
cpu: AMD EPYC 7763 64-Core Processor                
                          │   old.txt    │              new.txt               │
                          │    sec/op    │   sec/op     vs base               │
EvmInsertionSort-4          82.51m ± 10%   77.01m ± 3%  -6.67% (p=0.001 n=10)
EvmQuickSort-4              5.146m ±  1%   4.999m ± 1%  -2.86% (p=0.002 n=10)
EvmSignatureValidation-4    11.58µ ±  2%   11.50µ ± 1%       ~ (p=0.247 n=10)
EvmMulticallErcTransfer-4   4.765m ±  1%   4.729m ± 1%       ~ (p=0.063 n=10)
EvmRedBlackTree-4           225.6m ±  2%   223.6m ± 2%       ~ (p=0.123 n=10)
geomean                     5.554m         5.421m       -2.41%
```